check: split index/pack check into repository package
